% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemperatureData.R
\name{readTemperatureData}
\alias{readTemperatureData}
\title{Reads surface temperature data from a CSV file.}
\usage{
readTemperatureData(filePath)
}
\arguments{
\item{filePath}{String file name, either fully qualified or relative
to the current working directory.}
}
\value{
tibble with the data.
}
\description{
Reads surface temperature data from a CSV file.
}
