% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Produces a scatter plot of temperature anomalies by year.}
\usage{
scatterPlot()
}
\value{
ggplot object.
}
\description{
Produces a scatter plot of temperature anomalies by year.
}
